core/pagination.py
from django.conf import settings
from rest_framework.pagination import PageNumberPagination


class ResultSetPagination(PageNumberPagination):
    page_size = settings.REST_FRAMEWORK['PAGE_SIZE']
    page_size_query_param = 'page_size'
    max_page_size = settings.REST_FRAMEWORK['MAX_PAGE_SIZE']
