requirements.txt
# django
Django==4.0.4
python-decouple==3.6

# django rest framework
djangorestframework==3.13.1

# postgresql
psycopg2==2.9.3